moved glob.h to mega_glob.h
deleted glob.c and third_party folder
modified compilation files to point to the new location
