+++ examples/android/ExampleApp_ADT/jni/mega/Makefile.inc
@@ -57,8 +57,6 @@ CPP_SOURCES= sdk/src/attrmap.cpp \
     sdk/src/logging.cpp \
     sdk/src/mega_http_parser.cpp
     
-C_SOURCES = sdk/third_party/glob/glob.c
-
 HEADERS= sdk/include/mega.h \
 	    sdk/include/mega/account.h \
 	    sdk/include/mega/attrmap.h \
@@ -75,6 +73,7 @@ HEADERS= sdk/include/mega.h \
 	    sdk/include/mega/json.h \
 	    sdk/include/mega/megaapp.h \
 	    sdk/include/mega/megaclient.h \
+	    sdk/include/mega/mega_glob.h \
 	    sdk/include/mega/node.h \
 	    sdk/include/mega/pubkeyaction.h \
 	    sdk/include/mega/request.h \
@@ -100,12 +99,11 @@ HEADERS= sdk/include/mega.h \
         sdk/include/mega/posix/megafs.h  \
         sdk/include/mega/posix/megawaiter.h \
         sdk/include/mega/config-android.h \
-        sdk/third_party/glob/glob.h \
         sdk/include/mega/gfx/external.h \
         sdk/include/mega/thread/posixthread.h \
         sdk/include/mega/logging.h
 
-OBJECTS=$(CPP_SOURCES:.cpp=.o) $(C_SOURCES:.c=.o)
+OBJECTS=$(CPP_SOURCES:.cpp=.o)
 C_WRAPPER_SOURCES = sdk/src/megaapi.cpp sdk/src/megaapi_impl.cpp
 C_WRAPPER_OBJECTS = sdk/src/megaapi.o sdk/src/megaapi_impl.o
